Features now configured for compile-jsp server
