+++ liberty-maven-plugin/src/main/java/io/openliberty/tools/maven/BasicSupport.java
@@ -92,7 +92,6 @@ public class BasicSupport extends AbstractLibertySupport {
      */
     @Parameter(property = "serverName", defaultValue = "defaultServer")
     protected String serverName = null;
-    
     /**
      * Liberty user directory (<tT>WLP_USER_DIR</tt>).
      */
